wzsh: enable job control for spawned functions
I'm not sure exactly why I disabled that when I wrote this, but
it causes problems with "aliases".

For example, I use `vi() { vim "$@" }` in my startup to alias `vi` to
`vim`.

Without enabling job control, ctrl-z will background vim but then it
won't be able to be foregrounded.

Just enabling job control for this spawn seems to work, but I'm
not sure what other subtle issue this will cause because it's
been a while since I thought about this in depth.

Let's see what happens!